fin_tis_update

+++ app/Http/Controllers/Backend/FinishTissue/FinishTissueController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend\FinishTissue;
 use App\Http\Requests\Backend\FinishTissue\ManageFinishTissueRequest;
 use App\Http\Requests\Backend\FinishTissue\UpdateFinishTissueRequest;
 use App\Http\Requests\Backend\FinishTissue\StoreFinishTissueRequest;
+use App\Models\Collection\Collection;
 use App\Models\FinishTissue\FinishTissue;
 use App\Http\Controllers\Controller;
 use App\Repositories\Backend\FinishTissue\FinishTissueRepository;
@@ -45,7 +46,10 @@ class FinishTissueController extends Controller
      */
     public function create(ManageFinishTissueRequest $request)
     {
-        return view('backend.finishtissue.create');
+        $parents = FinishTissue::where('parent_id', null)->get();
+
+        return view('backend.finishtissue.create', [
+            'parents' => $parents]);
     }
 
     /**
@@ -55,24 +59,11 @@ class FinishTissueController extends Controller
      */
     public function store(StoreFinishTissueRequest $request)
     {
-        $this->finishTissue->create($request->only('type', 'title', 'title_ru', 'title_it'));
+        $this->finishTissue->create($request->all());
 
         $this->moveImg($request->photo);
 
         return redirect()->route('admin.finish-tissue.index')->withFlashSuccess(trans('alerts.backend.finishtissue.created'));
-    }
-
-    /**
-     *
-     * @return mixed
-     */
-    public function storeChild(FinishTissue $finishTissue)
-    {
-        FinishTissue::create([
-            'parent_id' => $finishTissue->id
-        ]);
-
-        return redirect()->route('admin.finish-tissue.edit', $finishTissue)->withFlashSuccess(trans('alerts.backend.finishtissue.created_child'));
     }
 
     /**
@@ -83,8 +74,11 @@ class FinishTissueController extends Controller
      */
     public function edit(FinishTissue $finishTissue, ManageFinishTissueRequest $request)
     {
+        $parents = FinishTissue::where('parent_id', null)->get();
+
         return view('backend.finishtissue.edit', [
             'finishTissue' => $finishTissue,
+            'parents' => $parents
         ]);
     }
 
@@ -96,21 +90,8 @@ class FinishTissueController extends Controller
      */
     public function update(FinishTissue $finishTissue, UpdateFinishTissueRequest $request)
     {
-        $children = $request['children'];
-
-        foreach ($children as $key => $newChild) {
-            $oldChild = FinishTissue::find($newChild['id']);
-            $oldImage = $oldChild->image;
-            $oldChild->parent_id = $finishTissue->id;
-            $oldChild->title = $newChild['title'];
-
-            $oldChild->image = $newChild['photo'];
-            if ($oldChild->save()) {
-                $this->moveImg($newChild['photo'], $oldImage);
-            }
-        }
         $oldName = $finishTissue->image;
-        $this->finishTissue->update($finishTissue, $request->only('type', 'title', 'title_ru', 'title_it'));
+        $this->finishTissue->update($finishTissue, $request->all());
         $this->moveImg($request->photo, $oldName);
 
         return redirect()->route('admin.finish-tissue.index')->withFlashSuccess(trans('alerts.backend.finishtissue.updated'));
@@ -125,6 +106,9 @@ class FinishTissueController extends Controller
     public function destroy(FinishTissue $finishTissue, ManageFinishTissueRequest $request)
     {
         $imgName = $finishTissue->image;
+        if($finishTissue->children()->count() > 0){
+            return redirect()->route('admin.finish-tissue.index')->withErrors(trans('exceptions.backend.access.finishtissue.delete_with_child_error'));
+        }
         $this->finishTissue->delete($finishTissue);
         $this->deleteImg($imgName);
